third_party/blink/web_tests/platform/generic/external/wpt/html/semantics/scripting-1/the-script-element/script-onerror-insertion-point-2-expected.txt

This is a testharness.js-based test.
FAIL Test that the insertion point is not defined in the error event of a
  parser-inserted script that has an unparseable URL assert_equals: expected "text" but got "Some text"
Harness: the test ran to completion.
